surya: Import servicetracker from common tree
* Blobs from LA.UM.9.14.r1-19300.01-LAHAINA.QSSI12.0

hwservicemanager: getTransport: Cannot find entry vendor.qti.hardware.servicetracker@1.0::IServicetracker/default in either framework or device VINTF manifest.

hwservicemanager: Since vendor.qti.hardware.servicetracker@1.0::IServicetracker/default is not registered, trying to start it as a lazy HAL.

libc: Unable to set property "ctl.interface_start" to "vendor.qti.hardware.servicetracker@1.0::IServicetracker/default": error code: 0x20

hwservicemanager: Tried to start vendor.qti.hardware.servicetracker@1.0::IServicetracker/default as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.

Signed-off-by: Richard Raya <rdxzv.dev@gmail.com>